finish chapter 7

+++ javascript/main.js
@@ -1,9 +1,15 @@
+if (localStorage.getItem("user-token") == null) {
+  window.location.replace(document.location.origin + "/login/");
+}
+
 function renderItems(items, processType, elementId, processFunction) {
   let placeholder = "<div>";
   let itemsMeta = [];
+
   for (i = 0; i < items.length; i++) {
     let title = items[i]["title"];
     let placeholderId = processType + "-" + title.replaceAll(" ", "-");
+
     placeholder +=
       '<div class="itemContainer">' +
       "<p>" +
@@ -20,6 +26,7 @@ function renderItems(items, processType, elementId, processFunction) {
   }
   placeholder += "</div>";
   document.getElementById(elementId).innerHTML = placeholder;
+
   for (i = 0; i < itemsMeta.length; i++) {
     document
       .getElementById(itemsMeta[i]["id"])
@@ -29,40 +36,41 @@ function renderItems(items, processType, elementId, processFunction) {
 
 function apiCall(url, method) {
   let xhr = new XMLHttpRequest();
-
   xhr.withCredentials = true;
   xhr.addEventListener("readystatechange", function () {
     if (this.readyState === this.DONE) {
-      renderItems(
-        JSON.parse(this.responseText)["pending_items"],
-        "edit",
-        "pendingItems",
-        editItem
-      );
-      renderItems(
-        JSON.parse(this.responseText)["done_items"],
-        "delete",
-        "doneItems",
-        deleteItem
-      );
+      if (this.status === 401) {
+        window.location.replace(document.location.origin + "/login/");
+      } else {
+        renderItems(
+          JSON.parse(this.responseText)["pending_items"],
+          "edit",
+          "pendingItems",
+          editItem
+        );
+        renderItems(
+          JSON.parse(this.responseText)["done_items"],
+          "delete",
+          "doneItems",
+          deleteItem
+        );
+        document.getElementById("completeNum").innerHTML = JSON.parse(
+          this.responseText
+        )["done_item_count"];
+        document.getElementById("pendingNum").innerHTML = JSON.parse(
+          this.responseText
+        )["pending_item_count"];
+      }
     }
-    document.getElementById("completeNum").innerHTML = JSON.parse(
-      this.responseText
-    )["done_item_count"];
-    document.getElementById("pendingNum").innerHTML = JSON.parse(
-      this.responseText
-    )["pending_item_count"];
   });
-
   xhr.open(method, url);
-  xhr.setRequestHeader("Content-Type", "application/json");
-  xhr.setRequestHeader("user-token", "token");
-
+  xhr.setRequestHeader("content-type", "application/json");
+  xhr.setRequestHeader("user-token", localStorage.getItem("user-token"));
   return xhr;
 }
 
 function editItem() {
-  let title = this.id.replaceAll("-", " ").replace("edit", "").trim();
+  let title = this.id.replaceAll("-", " ").replace("edit ", "");
   let call = apiCall("/item/edit", "PUT");
   let json = {
     title: title,
@@ -72,7 +80,7 @@ function editItem() {
 }
 
 function deleteItem() {
-  let title = this.id.replaceAll("-", " ").replace("delete", "").trim();
+  let title = this.id.replaceAll("-", " ").replace("delete ", "");
   let call = apiCall("/item/delete", "POST");
   let json = {
     title: title,
@@ -88,15 +96,11 @@ function getItems() {
 
 getItems();
 
+document.getElementById("create-button").addEventListener("click", createItem);
+
 function createItem() {
   let title = document.getElementById("name");
   let call = apiCall("/item/create/" + title.value, "POST");
   call.send();
   document.getElementById("name").value = null;
 }
-
-document.addEventListener("DOMContentLoaded", () => {
-  document
-    .getElementById("create-button")
-    .addEventListener("click", createItem);
-});
